# openapi.yaml
openapi: 3.0.3
info:
  title: logd API
  description: |
    分布式 PHP 日志中心 logd 的接口定义。

    测试顺序建议：
    1. 调用后台登录接口获取 logd_admin Cookie 和 csrf_token。
    2. 将 csrf_token 保存到接口调试软件环境中。
    3. 使用后台会话 Cookie 测试日志查询、Token、设置和状态接口。
    4. 使用创建 Token 返回的明文 Token 测试日志提交接口。
  version: 1.0.0
  contact:
    name: logd
servers:
  - url: http://127.0.0.1:7878
    description: 默认本地监听地址
...
paths:
  /api/v1/logs/batch:
    post:
      tags:
        - Ingest
      summary: 提交日志批次
      description: |
        PHP 站点使用项目 Token 提交日志批次。

        整个批次完整校验，任意一条日志不合法时整个请求返回 400，不支持部分成功。
        成功返回 202，只表示日志已经进入 logd 内存队列，不表示 PostgreSQL 已提交。
      operationId: ingestLogBatch
      security:
        - projectToken: []
      parameters:
        - name: Content-Encoding
          in: header
          required: false
          description: 设置为 gzip 时表示请求体为 gzip 压缩后的 JSON。
          schema:
            type: string
            enum:
              - gzip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogBatchRequest'
            example:
              node_key: node-sh-03
              logs:
                - event_id: 4f5b4fd5-42b8-47d5-a0db-0273cd76f574
                  event_time: '2026-09-21T12:30:00.123+08:00'
                  level: ERROR
                  request_ip: 1.2.3.4
                  member: '10001'
                  session_id: 2f3d9f38-1b3d-4e5c-9e74-9dc49b4f2fd0
                  request_method: POST
                  request_url: https://api.example.com/order/create
                  request_headers:
                    content-type: application/json
                    user-agent: PHP/8.3
                  request_params:
                    order_id: 10001
                  error_scene: payment.callback.signature_invalid
                  error_message: 支付回调验签失败
                  error_file: /www/shop/app/service/Payment.php
                  error_line: 126
                  error_stack: '...'
      responses:
        '202':
          description: 批次已进入内存队列
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestAccepted'
              example:
                accepted: 1
        '400':
          description: JSON、请求体或字段校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: INVALID_LOG
                  message: event_id is duplicated in batch
                  index: 1
                  field: event_id
        '401':
          description: 项目 Token 缺失或无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: UNAUTHORIZED
                  message: invalid bearer token
        '403':
          description: 项目已禁用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: PROJECT_DISABLED
                  message: project token is disabled
        '413':
          description: 解压后的请求体超过 5MB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: BODY_TOO_LARGE
                  message: decompressed request body exceeds 5MB
        '500':
          description: 服务内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: INTERNAL_ERROR
                  message: internal error
        '503':
          description: 接收队列已满
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: QUEUE_FULL
                  message: ingest queue is full
...
  /api/v1/admin/project-tokens:
    get:
      tags:
        - Admin Project Tokens
      summary: 获取项目 Token 列表
      operationId: listProjectTokens
      security:
        - adminSession: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTokenListResponse'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '500':
          description: 读取 Token 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: INTERNAL_ERROR
                  message: 读取 Token 失败
    post:
      tags:
        - Admin Project Tokens
      summary: 创建项目 Token
      description: Token 明文只在创建响应中返回一次。
      operationId: createProjectToken
      security:
        - adminSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectTokenRequest'
            example:
              project_key: shop-api
              name: 商城 API
              enabled: true
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectTokenResponse'
        '400':
          description: 请求或 Token 字段无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: INVALID_TOKEN
                  message: 项目标识长度必须在 1 到 64 个字符之间
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/AdminCSRFInvalid'
        '500':
          description: 创建 Token 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/admin/project-tokens/{id}:
    parameters:
      - $ref: '#/components/parameters/TokenID'
    patch:
      tags:
        - Admin Project Tokens
      summary: 修改项目 Token
      description: 只提交需要修改的字段。
      operationId: updateProjectToken
      security:
        - adminSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectTokenRequest'
            example:
              project_key: shop-api-v2
              name: 商城 API V2
              enabled: false
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectToken'
        '400':
          description: Token ID 或字段无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/AdminCSRFInvalid'
        '404':
          description: Token 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: NOT_FOUND
                  message: Token 不存在
        '500':
          description: 修改 Token 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags:
        - Admin Project Tokens
      summary: 删除项目 Token
      description: 删除为物理删除，删除后立即刷新内存缓存。
      operationId: deleteProjectToken
      security:
        - adminSession: []
          csrfToken: []
      responses:
        '204':
          description: 删除成功
        '400':
          description: Token ID 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: INVALID_TOKEN_ID
                  message: Token ID 无效
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/AdminCSRFInvalid'
        '404':
          description: Token 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: NOT_FOUND
                  message: Token 不存在
        '500':
          description: 删除 Token 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/admin/settings:
    get:
      tags:
        - Admin Settings
      summary: 获取系统设置
      operationId: getSettings
      security:
        - adminSession: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
              example:
                retention_days: 30
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '500':
          description: 读取系统设置失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags:
        - Admin Settings
      summary: 修改系统设置
      operationId: updateSettings
      security:
        - adminSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSystemSettingsRequest'
            example:
              retention_days: 30
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
        '400':
          description: 设置值无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: INVALID_SETTINGS
                  message: 保留天数必须在 1 到 3650 之间
        '401':
          $ref: '#/components/responses/AdminUnauthorized'
        '403':
          $ref: '#/components/responses/AdminCSRFInvalid'
        '500':
          description: 保存系统设置失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    projectToken:
      type: http
      scheme: bearer
      description: 项目 Token，通过后台创建后只在创建响应中返回一次。
    adminSession:
      type: apiKey
      in: cookie
      name: logd_admin
      description: 后台登录成功后由客户端保存的会话 Cookie。
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 后台登录响应返回的 csrf_token，POST、PATCH 和 DELETE 请求必须携带。
  parameters:
...
    TokenID:
      name: id
      in: path
      required: true
      description: 项目 Token 记录 UUID。
      schema:
        type: string
        format: uuid
  responses:
    AdminUnauthorized:
      description: 后台会话无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: admin session is invalid
    AdminCSRFInvalid:
      description: CSRF Token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: INVALID_CSRF
              message: CSRF token is invalid
  schemas:
    Level:
      type: string
      enum:
        - INFO
        - WARN
        - ERROR
        - DEBUG
    LogBatchRequest:
      type: object
      required:
        - node_key
        - logs
      properties:
        node_key:
          type: string
          minLength: 1
          maxLength: 64
          example: node-sh-03
        logs:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/LogItem'
    LogItem:
      type: object
      required:
        - event_id
        - event_time
        - level
        - request_ip
        - request_method
        - request_url
        - error_scene
        - error_message
      properties:
        event_id:
          type: string
          format: uuid
          description: 批次内不能重复。
          example: 4f5b4fd5-42b8-47d5-a0db-0273cd76f574
        event_time:
          type: string
          format: date-time
          description: 不能晚于服务当前时间 5 分钟。ERROR 最大允许延迟 31 天，其他级别使用 retention_days。
          example: '2026-09-21T12:30:00.123+08:00'
        level:
          $ref: '#/components/schemas/Level'
        request_ip:
          type: string
          description: 合法 IPv4 或 IPv6。
          example: 1.2.3.4
        member:
          type: string
          nullable: true
          maxLength: 255
          description: 可选用户信息，用于精确查询。
          example: '10001'
        session_id:
          type: string
          nullable: true
          maxLength: 255
          description: 可选 Session ID，用于精确查询。
          example: 2f3d9f38-1b3d-4e5c-9e74-9dc49b4f2fd0
        request_method:
          type: string
          minLength: 1
          maxLength: 16
          description: HTTP 请求方法，保存时转为大写。
          example: POST
        request_url:
          type: string
          minLength: 1
          maxLength: 8192
          example: https://api.example.com/order/create
        request_headers:
          type: object
          nullable: true
          additionalProperties: true
          description: JSON 对象，最大 64KB。
        request_params:
          type: object
          nullable: true
          additionalProperties: true
          description: JSON 对象，最大 64KB。
        error_scene:
          type: string
          minLength: 1
          maxLength: 255
          example: payment.callback.signature_invalid
        error_message:
          type: string
          minLength: 1
          maxLength: 16384
          example: 支付回调验签失败
        error_file:
          type: string
          nullable: true
          maxLength: 2048
          example: /www/shop/app/service/Payment.php
        error_line:
          type: integer
          nullable: true
          minimum: 1
          example: 126
        error_stack:
          type: string
          nullable: true
          maxLength: 65536
          example: '...'
    IngestAccepted:
      type: object
      required:
        - accepted
      properties:
        accepted:
          type: integer
          example: 1
...
    CreateProjectTokenRequest:
      type: object
      required:
        - project_key
        - name
      properties:
        project_key:
          type: string
          minLength: 1
          maxLength: 64
          example: shop-api
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: 商城 API
        enabled:
          type: boolean
          default: true
    UpdateProjectTokenRequest:
      type: object
      properties:
        project_key:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 100
        enabled:
          type: boolean
    ProjectToken:
      type: object
      required:
        - id
        - project_key
        - name
        - enabled
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        project_key:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectTokenResponse:
      type: object
      required:
        - token
        - project_token
      properties:
        token:
          type: string
          description: 只在创建响应中返回一次的 Token 明文。
          example: base64url-project-token
        project_token:
          $ref: '#/components/schemas/ProjectToken'
    ProjectTokenListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectToken'
    SystemSettings:
      type: object
      required:
        - retention_days
      properties:
        retention_days:
          type: integer
          minimum: 1
          maximum: 3650
          example: 30
    UpdateSystemSettingsRequest:
      type: object
      properties:
        retention_days:
          type: integer
          minimum: 1
          maximum: 3650
          example: 30
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: INVALID_LOG
        message:
          type: string
          example: event_time is invalid
        index:
          type: integer
          nullable: true
          description: 批次中的日志下标。
          example: 12
        field:
          type: string
          nullable: true
          example: event_time